+++ main.js
@@ -26,23 +26,34 @@ define(function (require, exports, module) {
        // var query = thisEditor._codeMirror.getSelection();
         NativeApp.openURLInDefaultBrowser("https://groups.google.com/forum/#!forum/brackets-dev");
     }
-
+    function Themes() {
+       // var thisEditor = EditorManager.getCurrentFullEditor();
+       // var query = thisEditor._codeMirror.getSelection();
+        NativeApp.openURLInDefaultBrowser("http://brackets-themes.github.io/");
+    }
 
     var MOZILLA_MDN = "Denisov.Mozilla";
     var W3_SCHOOLS = "Denisov.W3Schools";
     var FORUM_BRACKETS = "Denisov.Forum";
+    var BRACKETS_THEMES = "Denisov.Themes";
 
     CommandManager.register("Mozilla MDN", MOZILLA_MDN, Mozilla);
     CommandManager.register("W3 Schools Tutorial", W3_SCHOOLS, W3Shools);
-    CommandManager.register("Brackets-Dev", FORUM_BRACKETS, Forum);
+    CommandManager.register("Brackets-Dev (Google Group)", FORUM_BRACKETS, Forum);
+    CommandManager.register("Brackets Themes", BRACKETS_THEMES, Themes);
 
     var helpMenu = Menus.getMenu(Menus.AppMenuBar.HELP_MENU);
+    //var helpMenu = Menus.getMenu(Menus.AppMenuBar.HELP_MENU);
     //var contextMenu = Menus.getContextMenu(Menus.ContextMenuIds.EDITOR_MENU);*/
     helpMenu.addMenuDivider();
     helpMenu.addMenuItem(MOZILLA_MDN, "Ctrl-Alt-M");
     helpMenu.addMenuItem(W3_SCHOOLS, "Ctrl-Alt-W");
     helpMenu.addMenuItem(FORUM_BRACKETS, "Ctrl-Alt-B");
     helpMenu.addMenuDivider();
-    //contextMenu.addMenuDivider();
-    //contextMenu.addMenuItem(W3_SCHOOLS, "Ctrl-Alt-W");
-});
+
+
+    var viewMenu = Menus.getMenu(Menus.AppMenuBar.VIEW_MENU);
+    viewMenu.addMenuDivider();
+    viewMenu.addMenuItem(BRACKETS_THEMES, "Ctrl-Alt-H");
+
+ });